Continuing experiments with developing a page with photos
