+++ frontend/OpenArtspaceGallery/public/css/gallery.less
@@ -11,6 +11,7 @@
 @import "images/images.less";
 @import "images/thumbnail.less";
 @import "images/newImage.less";
+@import "images/photo";
 
 @import "shared/common.less";
 @import "shared/theme.less";
