+++ frontend/OpenArtspaceGallery/public/css/images/photo.less
@@ -1,0 +1,23 @@
+@import "../shared/common.less";
+@import "../shared/theme";
+
+.tool-bar {
+  background-color: green;
+  height: 2rem;
+}
+
+.image-section {
+  flex: 1;
+  background-color: red;
+  height: 10rem;
+}
+
+.info-section {
+  background-color: blue;
+  width: 33%;
+  height: 33%;
+}
+
+.image-content-container {
+  display: flex;
+}
