--- frontend/OpenArtspaceGallery/src/views/AdminPanelView.vue
@@ -1,15 +1,0 @@
-<script setup lang="ts">
-  import HeaderComponent from "../components/Shared/Header/HeaderComponent.vue";
-  import FooterComponent from "../components/Shared/Foooter/FooterComponent.vue";
-  import AdminPanelComponent from "../components/AdminPanel/AdminPanelComponent.vue";
-</script>
-
-<template>
-
-  <HeaderComponent />
-
-  <AdminPanelComponent />
-
-  <FooterComponent/>
-
-</template>
