--- frontend/OpenArtspaceGallery/src/views/AlbumContentView.vue
@@ -1,17 +1,0 @@
-<script setup lang="ts">
-  import FooterComponent from "../components/Shared/Foooter/FooterComponent.vue";
-  import HeaderComponent from "../components/Shared/Header/HeaderComponent.vue";
-  import AlbumPageComponent from "../components/Albums/AlbumPageComponent.vue";
-
-  const props = defineProps({
-    currentAlbumId: String
-  })
-</script>
-
-<template>
-  <HeaderComponent />
-
-  <AlbumPageComponent :currentAlbumId="props.currentAlbumId" />
-
-  <FooterComponent />
-</template>
